requirements.txt
pandas
selenium
pytest
pytest-html
openpyxl
python-dotenv